% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide_moloc.R
\name{coloc.eqtl.biom.mqtl}
\alias{coloc.eqtl.biom.mqtl}
\title{coloc.eqtl.biom}
\usage{
coloc.eqtl.biom.mqtl(listData, bed, outfolder = "test", prefix = "pref",
  save.SNP.info = FALSE, cores = 20, have_alleles = TRUE,
  bychrpos = TRUE, prior_var = c(0.01, 0.1, 0.5), priors = c(0.0001,
  1e-06, 1e-07), min_nsnps = 50)
}
\arguments{
\item{listData, }{a list of data frames in this order: gwas, eqtl, mqtl
The data frames need to have columns: "SNP" or "CHR", "POS"; "BETA", "SE"; 
                                       If eqtl/mqtl: "ProbeID"
                                       Optionally, "A1", "A2" if want to match alleles}

\item{bed, }{a data frame with "CHR", "START", "STOP", and 
"ProbeID" matching either eqtl or mqtl}

\item{outfolder, }{path of folder where the results will be written to}

\item{prefix, }{character, name to give to the output file}

\item{save.SNP.info, }{logical, should info of each SNP be saved? 
If this option is true, beware that it will takes up a lot of time and space.}

\item{cores, }{number. See foreach package.}

\item{have_alleles, }{logical. If TRUE, matches alleles to be the same as gwas data.}

\item{bychrpos, }{logical. If TRUE, uses "CHR", "POS" columns 
as SNP id to match across data frames;
If TRUE, make sure not to have another column called "SNP".}

\item{prior_var, }{numeric vector specifying any number of values;
These will be used to average the ABF computation across different prior variances.}

\item{priors, }{numeric vectors with three numbers; 
First number is the prior of associaion of a SNP with any of the traits;
Second number is the prior of a SNP being associated with two traits;
Third number is the prior of a SNP being associated with three traits.}

\item{min_nsnps, }{number. The minimum number of matching SNPs across 
the three data frames to consider.}
}
\value{
results in outfolder: text file with colocalization results for each locus; 
        text file with removed snps; 
        if save.SNP.info = TRUE, a folder called "coloc.output.perSNP"
        will save a file per locus with SNP information such as ABF (one row per matching SNP)
}
\description{
Genome-wide co-localization with three traits
}
\examples{
listData = list(gwas=biom.df, eqtl=eqtl.df, mqtl=methyl.df)

}
\author{
Claudia Giambartolomei
}

