% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moloc.R
\name{moloc_test}
\alias{moloc_test}
\title{moloc_test}
\usage{
moloc_test(listData, overlap = FALSE, prior_var = c(0.01, 0.1, 0.5),
  priors = c(0.0001, 1e-06, 1e-07), from_p = FALSE)
}
\arguments{
\item{listData}{A list of data frames to be analyzed. Each data frame
must contain columns named SNP (the SNP name consistent across all the data frames), 
BETA, 
SE}

\item{overlap}{Logical, do the individuals in the datasets overlap?}

\item{prior_var}{One or more numbers for the prior variance of the ABF.}

\item{priors}{are numbers, the prior for one variant to be associated with 1 trait and with each additional trait}

\item{...}{parameters passed to \code{\link{adjust_bfs}}, \code{\link{config_coloc}}, \code{\link{snp_ppa}}}
}
\value{
A list of three elements: 
        First is a data frame with 4 variables: priors ('prior'), likelihoods ('sumbf') and 
          posteriors ('loglkl' and 'PPA') for each configuration; 
        Second is a number, the number of SNPs in common in the region;
        Third is a data frame with 2 variables: 
          SNP with the best posterior for each scenario ('coloc_ppas'), 
          SNP name with the best posterior for each scenario ('best.snp.coloc').
}
\description{
Bayesian multiple trait colocalization analysis using list of data.frames
}
\details{
Runs \code{\link{adjust_bfs}}, \code{\link{config_coloc}}, \code{\link{snp_ppa}}
}
\examples{
moloc <- moloc_test(listData) # uses default priors

}
\author{
Claudia Giambartolomei

Claudia Giambartolomei
}

